requirements.txt
Django>=3.2,<4
djangorestframework>=3.9,<4
gdal>=1.6
iso8601
htimeseries>=3.1,<4
pytz
rules>=2.0
dj-rest-auth==2.1.7
django-rest-captcha>=0.1.0,<1
django-parler>=2.2,<3
django-parler-rest>=2,<3
django-geowidgets>=0.1.1,<1
django-nested-admin>=3.2,<4
django-crequest
celery>=5,<6
django-registration-redux>=2.9,<3
django-simple-captcha>=0.5.14,<1
django-bootstrap4>=2,<4
requests>=2.25,<3
